+++ queries/qml/imports.scm
@@ -1,0 +1,3 @@
+(ui_import source: (identifier) @module)
+(ui_import source: (nested_identifier) @module)
+(ui_import source: (string) @module)